moved components under the lib directory
