+++ src/lib/components/game/modals/GameEnded.svelte
@@ -1,22 +1,19 @@
-<script>
-	import { player, resign } from '../../../stores';
+<script lang="ts">
+	import { winner, player } from '../../../../stores';
 </script>
 
 <div class="container">
 	<div class="wrap">
-		{#if $resign.resigner !== $player}
-			Your opponent has resigned the game.
+		{#if $winner == $player}
+			Congratulations! You have won the game.
+			<img src="/win.png" class="icon" />
 		{:else}
-			You have resigned the game.
+			You lost, the {$winner} was victorious.
+			<img src="/lose.png" class="icon" />
 		{/if}
-		<img src="/resign.png" class="icon" />
-		<div class="linkbox">
-			<a href="/game" on:click={() => ($resign = { resigned: false, resigner: '' })}
-				>Start a new game</a
-			>
-			<a href="/profile" on:click={() => ($resign = { resigned: false, resigner: '' })}
-				>Go to profile</a
-			>
+		<div>
+			<a href="/game" on:click={() => ($winner = '')}>Start a new game</a>
+			<a href="/profile on:click={() => ($winner = '')}">Go to profile</a>
 		</div>
 	</div>
 </div>
@@ -38,7 +35,6 @@
 		padding: 2rem;
 		padding-inline: 4rem;
 		border: lightgray solid 0.5px;
-		border-radius: 5px;
 		box-shadow: rgba(50, 50, 93, 0.25) 0px 30px 60px -12px inset,
 			rgba(0, 0, 0, 0.3) 0px 18px 36px -18px inset;
 		color: white;
@@ -47,16 +43,13 @@
 		align-items: center;
 		gap: 0.5rem;
 	}
-	.linkbox {
-		margin-top: 1rem;
-	}
-	a {
-		color: grey;
-		margin: 1rem;
-	}
 	.icon {
 		height: 2rem;
 		width: auto;
-		margin-top: 1rem;
+		margin: 1rem;
+	}
+	a {
+		margin: 1rem;
+		color: lightgrey;
 	}
 </style>
