+++ src/lib/components/profile/Stats.svelte
